requirements-frozen.txt
# Generated by: pip freeze
# Python 3.13
# Reproduces the exact environment — install with: pip install -r requirements-frozen.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arize-phoenix-otel==0.15.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
coverage==7.13.5
fastapi==0.135.3
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.4
openinference-instrumentation==0.1.46
openinference-semantic-conventions==0.1.28
openrouter==0.8.1
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-asgi==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
packaging==26.0
pandas==3.0.2
pluggy==1.6.0
protobuf==6.33.6
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
six==1.17.0
sql_metadata==2.20.0
sqlglot==30.4.2
sqlparse==0.5.5
starlette==1.0.0
structlog==25.5.0
tenacity==9.1.4
tiktoken==0.12.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
zipp==3.23.1
